Regroup non-golden tests by concern instead of by input format
Moves test functions across files so each ends up in the directory
that matches its actual concern, not the input format it happens to
use:

- Consolidate R/Dhall/Nix dict-key error tests, previously duplicated
  across tests/formats/test_json.py and tests/formats/test_yaml.py,
  into a single tests/errors/test_dict_key_errors.py parameterized
  over format.
- Drop redundant mixed-dict-values tests from tests/formats/test_json.py
  and tests/formats/test_yaml.py; tests/errors/test_coercion_errors.py
  already covers them across all formats.
- Keep YAML-tag-specific heterogeneous checks (!!binary, !!omap, !!set,
  native date/datetime) in tests/formats/test_yaml.py where they belong.
- Move positive-path tests out of tests/errors/test_call_errors.py
  (Haskell ref-strip behaviors, Python dotted call success, Python
  camel-case ref acceptance, Raku kebab ref support, Python variable
  names support, Dhall quoted dict key) into tests/test_languages.py.
- Move the supported_ref_cases / identifier_cases invariant check to
  tests/metadata/test_language_metadata.py.
